Fix integer overflow in GenBank release file size parsing
identify_downloadable_files() parsed each file's byte size with
as.integer(), which silently returns NA for values above the 32-bit
integer limit (~2.1 billion). This release has individual .seq files
over 2 GB, so 72 files' sizes were becoming NA, triggering spurious
"NAs introduced by coercion" and "Not all file information could be
ascertained" warnings from get_plants_index(). The filesizes column
isn't otherwise used, so this had no effect on the actual result, but
switching to as.numeric() fixes it at the source.

+++ R/setup_gb_functions.R
@@ -84,7 +84,7 @@ identify_downloadable_files <- function(flpth) {
                    x = descripts)
   filesize_info <- strsplit(x = filesize_lines, split = '\\s')
   filesize_info <- lapply(X = filesize_info, function(x) x[x != ''])
-  filesizes <- as.integer(vapply(X = filesize_info, FUN = '[[', i = 1,
+  filesizes <- as.numeric(vapply(X = filesize_info, FUN = '[[', i = 1,
                                  FUN.VALUE = character(1)))
   names(filesizes) <- vapply(X = filesize_info, FUN = '[[', i = 2,
                              FUN.VALUE = character(1))
